frontend/requirements.txt
Flask
Werkzeug
Pillow
requests

